Componentize the staggered transition group
